requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
billiard==4.2.0
boto3==1.35.17
botocore==1.35.17
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.1
fastapi-cli==0.0.5
flower==2.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
humanize==4.10.0
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
openai==1.45.0
openapi==2.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
